fix: unit tests
